refactor(guest-agent): extract codex setup boundary
